Compute embedding for all images in competition
Given this is not a code competition, there will
be no additional samples to the ones already provided.
When a model is trained, just use it to compute the
embeddings for all of these provided samples.

These embeddings can then be further used for different
purposes, like local validation, or submission.  These
all use retrieval via embedding distances.